Finished moving data to JSON
