+++ tests/test_data/data/2018-19/gamerotation/data_00218DUMMY1.json
@@ -1,0 +1,112 @@
+{
+    "resultSets": [
+        {
+            "name": "AwayTeam",
+            "headers": [
+                "GAME_ID",
+                "TEAM_ID",
+                "PERSON_ID",
+                "IN_TIME_REAL",
+                "OUT_TIME_REAL"
+            ],
+            "rowSet": [
+                [
+                    "00218DUMMY1",
+                    1610612760,
+                    7,
+                    0.0,
+                    28800.0
+                ],
+                [
+                    "00218DUMMY1",
+                    1610612760,
+                    8,
+                    0.0,
+                    28800.0
+                ],
+                [
+                    "00218DUMMY1",
+                    1610612760,
+                    9,
+                    0.0,
+                    200.0
+                ],
+                [
+                    "00218DUMMY1",
+                    1610612760,
+                    10,
+                    0.0,
+                    28800.0
+                ],
+                [
+                    "00218DUMMY1",
+                    1610612760,
+                    11,
+                    0.0,
+                    28800.0
+                ],
+                [
+                    "00218DUMMY1",
+                    1610612760,
+                    12,
+                    200.0,
+                    28800.0
+                ]
+            ]
+        },
+        {
+            "name": "HomeTeam",
+            "headers": [
+                "GAME_ID",
+                "TEAM_ID",
+                "PERSON_ID",
+                "IN_TIME_REAL",
+                "OUT_TIME_REAL"
+            ],
+            "rowSet": [
+                [
+                    "00218DUMMY1",
+                    1610612761,
+                    1,
+                    0.0,
+                    28800.0
+                ],
+                [
+                    "00218DUMMY1",
+                    1610612761,
+                    2,
+                    0.0,
+                    28800.0
+                ],
+                [
+                    "00218DUMMY1",
+                    1610612761,
+                    3,
+                    0.0,
+                    28800.0
+                ],
+                [
+                    "00218DUMMY1",
+                    1610612761,
+                    4,
+                    0.0,
+                    28800.0
+                ],
+                [
+                    "00218DUMMY1",
+                    1610612761,
+                    5,
+                    0.0,
+                    200.0
+                ],
+                [
+                    "00218DUMMY1",
+                    1610612761,
+                    6,
+                    200.0,
+                    28800.0
+                ]
+            ]
+        }
+    ]
+}
