+++ tests/test_data/data/2018-19/playerdashboardbygeneralsplits/data_7.json
@@ -1,0 +1,45 @@
+{
+    "resultSets": [
+        {
+            "name": "OverallPlayerDashboard",
+            "headers": [
+                "FT_PCT"
+            ],
+            "rowSet": [
+                [
+                    0.85
+                ]
+            ]
+        },
+        {
+            "name": "LocationPlayerDashboard",
+            "headers": [],
+            "rowSet": []
+        },
+        {
+            "name": "WinsLossesPlayerDashboard",
+            "headers": [],
+            "rowSet": []
+        },
+        {
+            "name": "MonthPlayerDashboard",
+            "headers": [],
+            "rowSet": []
+        },
+        {
+            "name": "PrePostAllStarPlayerDashboard",
+            "headers": [],
+            "rowSet": []
+        },
+        {
+            "name": "StartingPosition",
+            "headers": [],
+            "rowSet": []
+        },
+        {
+            "name": "DaysRestPlayerDashboard",
+            "headers": [],
+            "rowSet": []
+        }
+    ]
+}
